requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.7
# by the following command:
#
#    make reqs
#
ansible-runner==2.3.1
    # via -r requirements.in
docutils==0.19
    # via python-daemon
lockfile==0.12.2
    # via python-daemon
packaging==23.0
    # via ansible-runner
pexpect==4.8.0
    # via ansible-runner
ptyprocess==0.7.0
    # via pexpect
python-daemon==2.3.2
    # via ansible-runner
python-retry-decorator @ git+https://github.com/reyalsorik/python-retry-decorator
    # via -r requirements.in
pyyaml==6.0
    # via ansible-runner
six==1.16.0
    # via ansible-runner
tenacity==8.1.0
    # via python-retry-decorator

# The following packages are considered to be unsafe in a requirements file:
# setuptools
